Fixes problems with zero weight loops.
Zero weight loops make graph contraction quite complicated. When we contract
a node and thereby calculate a shortest path between the source and target
edges we always need to decide if the resulting path is a bridge path
(-> we need a shortcut) or not. In many cases such a path is not a bridge
path, because it is only an extension of another path that is a bridge path.
A special case is a path that goes to the target node, takes a loop
and then goes to the target edge. Normally we would not need a shortcut
for such a path, but when there are two or more such loops with zero weight
we have to be very careful not to miss a shortcut. For example it can happen
that there is no shortcut for one of the loops (as target edge) because
of the other loop and vice versa.